[BUGFIX] Fix all psalm issues

+++ Classes/Service/Database/UidReservationService.php
@@ -58,7 +58,7 @@ class UidReservationService
      */
     public function getReservedUid(): int
     {
-        $nextAutoIncrement = (int)max(
+        $nextAutoIncrement = max(
             $this->fetchSysFileAutoIncrementFromDatabase($this->localDatabaseConnection),
             $this->fetchSysFileAutoIncrementFromDatabase($this->foreignDatabaseConnection)
         );
